% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversations.R
\name{all_conversation_replies}
\alias{all_conversation_replies}
\title{Retrieve All Replies}
\usage{
all_conversation_replies(
  conversations,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN"),
  max_results = Inf,
  max_calls = Inf,
  limit = 1000L
)
}
\arguments{
\item{conversations}{The list returned by \code{\link{conversations}}.}

\item{...}{Arguments to pass to method}

\item{token}{Character. Your Slack API token. Default: Sys.getenv("SLACK_API_TOKEN")}

\item{max_results}{Integer. The maximum number of results to return (total).
Note: The actual maximum will be \code{max_results} rounded up to a
multiple of \code{limit}. Default: Inf}

\item{max_calls}{Integer. The maximum number of separate API calls to make
while constructing the response. Default: Inf}

\item{limit}{Numeric, maximum number of results per call. Default: 1000L}
}
\value{
A nested list with additional class "channel.replies". The return has
the same length as conversations. Each element of the list is an individual
return from \code{\link{conversation_replies}}.
}
\description{
Retrieve replies to all conversations from a channel on a Slack
team.
}
