% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversations.R
\name{conversations}
\alias{conversations}
\alias{replies}
\title{Retrieve Conversations and Replies}
\usage{
conversations(channel, ..., api_token = Sys.getenv("SLACK_API_TOKEN"),
  max_results = Inf, max_calls = Inf, paginate = TRUE)

replies(ts, channel, ..., api_token = Sys.getenv("SLACK_API_TOKEN"),
  max_results = Inf, max_calls = Inf, paginate = TRUE)
}
\arguments{
\item{channel}{character, Channel name to retieve conversation from}

\item{...}{arguments to pass to method}

\item{api_token}{character, full Slack API token, Default: Sys.getenv("SLACK_API_TOKEN")}

\item{ts}{character, Unique identifier of a thread's parent message}
}
\value{
A \link[httr:response]{response} object containing channel messages with the class "conversation"
}
\description{
Retrieve Conversations from a Channel on a Slack Team
}
\details{
To use the method, you'll need at least one of the channels,
groups, im or mpim scopes corresponding to the conversation
type you're working with.

See \href{https://api.slack.com/methods/conversations.history}{Slack Documentation} for more details on the arguments that the method can recieve in \dots.
}
