% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replies.R
\name{get_replies}
\alias{get_replies}
\title{FUNCTION_TITLE}
\usage{
get_replies(ts, ..., channel = Sys.getenv("SLACK_CHANNEL"),
  api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{ts}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{channel}{PARAM_DESCRIPTION, Default: Sys.getenv("SLACK_CHANNEL")}

\item{api_token}{PARAM_DESCRIPTION, Default: Sys.getenv("SLACK_API_TOKEN")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[slackteams:get_team_channels]{get_team_channels}
\link[httr:POST]{POST}, \link[httr:warn_for_status]{warn_for_status}, \link[httr:content]{content}
}
