% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{paginate}
\alias{paginate}
\title{Pagination}
\usage{
paginate(parent, ...)
}
\arguments{
\item{parent}{conversation or reply class object}

\item{...}{arguments to pass to get_conversation or get_reply}
}
\value{
list of \link[httr:response]{response} objects containing channel messages or thread messages
}
\description{
Paginating through collections of messages and replies.
}
\details{
See \href{https://api.slack.com/docs/pagination}{Slack Documentation} for more details.
}
