% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{paginate.reply}
\alias{paginate.reply}
\title{FUNCTION_TITLE}
\usage{
paginate.reply(root, ...)
}
\arguments{
\item{root}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
