% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversations.R
\name{conversations}
\alias{conversations}
\title{Retrieve Conversations}
\usage{
conversations(channel, ..., token = Sys.getenv("SLACK_API_TOKEN"),
  max_results = Inf, max_calls = Inf, limit = 1000L)
}
\arguments{
\item{channel}{Character. Channel name to retrieve conversation from.}

\item{...}{Arguments to pass to method}

\item{token}{Character. Your Slack API token. Default: Sys.getenv("SLACK_API_TOKEN")}

\item{max_results}{Integer. The maximum number of results to return (total).
Note: The actual maximum will be \code{max_results} rounded up to a
multiple of \code{limit}. Default: Inf}

\item{max_calls}{Integer. The maximum number of separate API calls to make
while constructing the response. Default: Inf}

\item{limit}{Numeric, maximum number of results per call. Default: 1000L}
}
\value{
A list of channel messages with class "conversations.history".
}
\description{
Retrieve conversations from a channel on a Slack team.
}
\details{
To use the method, you'll need at least one of the channels, groups,
im or mpim scopes corresponding to the conversation type you're working
with.

See \href{https://api.slack.com/methods/conversations.history}{Slack Documentation} for
more details on the arguments that the method can recieve in \dots.
}
