% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversations.R
\name{get_permalink}
\alias{get_permalink}
\title{POST Message Permalink}
\usage{
get_permalink(
  channel = Sys.getenv("SLACK_CHANNEL"),
  api_token = Sys.getenv("SLACK_API_TOKEN"),
  timestamp
)
}
\arguments{
\item{channel}{character, Conversation ID to fetch history for, Default: Sys.getenv("SLACK_CHANNEL")}

\item{api_token}{character, full Slack API token, Default: Sys.getenv("SLACK_API_TOKEN")}

\item{timestamp}{Character (or coercible to character); the timestamp of a
particular message.}
}
\value{
Character; the permalink to that particular message.
}
\description{
Individual messages are indexed by channel and timestamp. This API method
returns a url for the particular message.
}
