% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replies.R
\name{get_replies}
\alias{get_replies}
\title{GET Conversation Thread}
\usage{
get_replies(ts, ..., channel = Sys.getenv("SLACK_CHANNEL"),
  api_token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{ts}{character, Unique identifier of a thread's parent message}

\item{...}{arguments to pass to GET call}

\item{channel}{character, channel name, Default: Sys.getenv("SLACK_CHANNEL")}

\item{api_token}{character, full Slack API token, Default: Sys.getenv("SLACK_API_TOKEN")}
}
\value{
A \link[httr:response]{response} object containing thread messages with the class "reply"
}
\description{
Conversations API method returns an entire thread
(a message plus all the messages in reply to it),
while conversations.history method returns only parent messages.
}
\details{
See \href{https://api.slack.com/methods/conversations.replies}{Slack Documentation} for more details.
}
\seealso{
\link[slackteams:get_team_channels]{get_team_channels}
\link[httr:POST]{POST}, \link[httr:warn_for_status]{warn_for_status}, \link[httr:content]{content}
}
