% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversations.R
\name{replies}
\alias{replies}
\alias{conversation_replies}
\title{Retrieve Replies}
\usage{
replies(
  ts,
  channel,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN"),
  max_results = Inf,
  max_calls = Inf,
  limit = 1000L
)

conversation_replies(
  conversation,
  ...,
  token = Sys.getenv("SLACK_API_TOKEN"),
  max_results = Inf,
  max_calls = Inf,
  limit = 1000L
)
}
\arguments{
\item{ts}{Character. Unique identifier of a thread's parent message.}

\item{channel}{Character. Channel name to retrieve conversation from.}

\item{...}{Arguments to pass to method}

\item{token}{Character. Your Slack API token. Default: Sys.getenv("SLACK_API_TOKEN")}

\item{max_results}{Integer. The maximum number of results to return (total).
Note: The actual maximum will be \code{max_results} rounded up to a
multiple of \code{limit}. Default: Inf}

\item{max_calls}{Integer. The maximum number of separate API calls to make
while constructing the response. Default: Inf}

\item{limit}{Numeric, maximum number of results per call. Default: 1000L}

\item{conversation}{An individual conversation object from the list returned
by \code{\link{conversations}}.}
}
\description{
Retrieve replies to individual conversations on a Slack team.
}
